+++ NEWS.md
@@ -1,5 +1,6 @@
 ## v1.0.9
 
+* `phe_life_expectancy()` function added which allows users to calculate life expectancy at different ages based on population and death data
 
 ## V1.0.8
 This is the first release of this package to CRAN